pyproject.toml
[tool.poetry]
name = "effectivemobiletesttask"
version = "0.1.0"
description = ""
authors = ["Pechenuxa1 <v.tatarintsev@g.nsu.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.121.2"
uvicorn = "^0.38.0"
dotenv = "^0.9.9"
bcrypt = "^5.0.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.44"}
pyjwt = "^2.10.1"
redis = "^7.0.1"
asyncpg = "^0.30.0"
psycopg2-binary = "^2.9.11"
pydantic = {extras = ["email"], version = "^2.12.4"}
alembic = "^1.17.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
